requirements.txt
openpyxl~=3.1.2
pandas~=2.0.3
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
httplib2~=0.22.0
oauthlib~=3.2.2